add SmallDataAna_psana base class. Clean up and get ready to make the LCLS-II version"
